\name{tsq-package}
\alias{tsq-package}
\alias{tsq}
\docType{package}
\title{
\packageTitle{tsq}
}
\description{
\packageDescription{tsq}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{tsq}
\packageIndices{tsq}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{tsq}

Maintainer: \packageMaintainer{tsq}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
